AppFrame.tsx App wrapper, prevent rendering map twice when at routes page
