+++ src/components/appFrame.scss
@@ -1,0 +1,6 @@
+.appContent {
+    display: flex;
+    flex-direction: row-reverse;
+    height: 100%;
+    overflow-y: auto;
+}
