+++ src/components/map/Map.tsx
@@ -2,7 +2,7 @@ import classnames from 'classnames';
 import * as L from 'leaflet';
 import 'leaflet-editable';
 import 'leaflet/dist/leaflet.css';
-import { reaction, toJS, IReactionDisposer } from 'mobx';
+import { reaction, IReactionDisposer } from 'mobx';
 import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { LayerContainer, Map, Pane, TileLayer, ZoomControl } from 'react-leaflet';
@@ -139,7 +139,6 @@ class LeafletMap extends React.Component<IMapProps> {
 
     render() {
         const isMapInteractionRestricted = this.props.mapStore!.isMapInteractionRestricted;
-        const routes = toJS(this.props.routeListStore!.routes);
         return (
             <div className={s.mapView} data-cy='mapView'>
                 {this.props.children}
@@ -168,7 +167,7 @@ class LeafletMap extends React.Component<IMapProps> {
                     <NetworkLayers />
                     <EditNodeLayer />
                     <EditLinkLayer />
-                    <RouteLayer routes={routes} />
+                    <RouteLayer />
                     <EditRoutePathLayer />
                     <PopupLayer />
                     <StopAreaLayer />
